ER-290 - Deleting SkillsService and moving logic into SkillsOrchestrator as it does not need to be reused
