+++ src/Employer/Employer.Web/Configuration/IoC.cs
@@ -36,7 +36,6 @@ namespace Esfa.Recruit.Employer.Web.Configuration
         {
             services.AddTransient<ITrainingProviderService, TrainingProviderService>();
             services.AddTransient<IGeocodeImageService>(_ => new GoogleMapsGeocodeImageService(configuration.GetValue<string>("GoogleMapsPrivateKey")));
-            services.AddSingleton<ISkillsService, SkillsService>();
         }
 
         private static void RegisterProviderApiClientDep(IServiceCollection services, IConfiguration configuration)
