+++ src/Employer/Employer.Web/Orchestrators/Part2/SkillsOrchestrator.cs
@@ -2,24 +2,24 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Esfa.Recruit.Employer.Web.Services;
+using Esfa.Recruit.Employer.Web.Configuration;
 using Esfa.Recruit.Employer.Web.ViewModels.Part2.Skills;
 using Esfa.Recruit.Vacancies.Client.Domain;
 using Esfa.Recruit.Vacancies.Client.Domain.Exceptions;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.Client;
-using Microsoft.AspNetCore.Mvc.ViewFeatures.Internal;
+using Microsoft.Extensions.Options;
 
 namespace Esfa.Recruit.Employer.Web.Orchestrators.Part2
 {
     public class SkillsOrchestrator
     {
         private readonly IVacancyClient _client;
-        private readonly ISkillsService _skillsService;
-        
-        public SkillsOrchestrator(IVacancyClient client, ISkillsService skillsService)
+        private readonly SkillsConfiguration _skillsConfig;
+
+        public SkillsOrchestrator(IVacancyClient client, IOptions<SkillsConfiguration> skillsConfigOptions)
         {
             _client = client;
-            _skillsService = skillsService;
+            _skillsConfig = skillsConfigOptions.Value;
         }
 
         public async Task<SkillsViewModel> GetSkillsViewModelAsync(Guid vacancyId)
@@ -51,9 +51,9 @@ namespace Esfa.Recruit.Employer.Web.Orchestrators.Part2
 
         public void SetViewModelSkills(SkillsViewModel vm, IEnumerable<string> skills)
         {
-            vm.Column1Checkboxes = _skillsService.GetColumn1ViewModel(skills).ToList();
-            vm.Column2Checkboxes = _skillsService.GetColumn2ViewModel(skills).ToList();
-            vm.CustomSkills = _skillsService.GetCustomSkills(skills).ToList();
+            vm.Column1Checkboxes = GetColumn1ViewModel(skills).ToList();
+            vm.Column2Checkboxes = GetColumn2ViewModel(skills).ToList();
+            vm.CustomSkills = GetCustomSkills(skills).ToList();
         }
 
         public async Task PostSkillsEditModelAsync(SkillsEditModel m)
@@ -67,9 +67,49 @@ namespace Esfa.Recruit.Employer.Web.Orchestrators.Part2
             
             var skills = m.Skills ?? new List<string>();
             
-            vacancy.Skills = _skillsService.SortSkills(skills).ToList();
+            vacancy.Skills = SortSkills(skills).ToList();
 
             await _client.UpdateVacancyAsync(vacancy, false);
+        }
+
+        private IEnumerable<SkillViewModel> GetColumn1ViewModel(IEnumerable<string> selected)
+        {
+            return _skillsConfig.Column1Skills.Select(c => new SkillViewModel
+            {
+                Name = c,
+                Selected = selected != null && selected.Any(s => s == c)
+            });
+        }
+
+        private IEnumerable<SkillViewModel> GetColumn2ViewModel(IEnumerable<string> selected)
+        {
+            return _skillsConfig.Column2Skills.Select(c => new SkillViewModel
+            {
+                Name = c,
+                Selected = selected != null && selected.Any(s => s == c)
+            });
+        }
+
+        private IEnumerable<string> GetCustomSkills(IEnumerable<string> selected)
+        {
+            if (selected == null)
+            {
+                return new List<string>();
+            }
+
+            return selected.Except(_skillsConfig.Column1Skills).Except(_skillsConfig.Column2Skills);
+        }
+
+        private IEnumerable<string> SortSkills(IEnumerable<string> selected)
+        {
+            var filteredSelectedSkills = selected.Where(s => !string.IsNullOrWhiteSpace(s)).Distinct();
+
+            var orderedSkills = _skillsConfig.Column1Skills
+                .Union(_skillsConfig.Column2Skills)
+                .Union(selected)
+                .ToList();
+
+            return orderedSkills.Where(filteredSelectedSkills.Contains);
         }
     }
 }
